move JDBCBackendHandler.getConnection => ConnectionManager
