+++ sharding-proxy/src/main/java/io/shardingsphere/proxy/frontend/mysql/MySQLFrontendHandler.java
@@ -20,7 +20,7 @@ package io.shardingsphere.proxy.frontend.mysql;
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.ChannelHandlerContext;
 import io.shardingsphere.core.routing.router.masterslave.MasterVisitedManager;
-import io.shardingsphere.proxy.backend.common.ProxyConnectionHolder;
+import io.shardingsphere.proxy.backend.common.jdbc.ConnectionManager;
 import io.shardingsphere.proxy.frontend.common.FrontendHandler;
 import io.shardingsphere.proxy.frontend.common.executor.ExecutorGroup;
 import io.shardingsphere.proxy.transport.common.packet.DatabaseProtocolPacket;
@@ -92,7 +92,7 @@ public final class MySQLFrontendHandler extends FrontendHandler {
                     }
                 } finally {
                     MasterVisitedManager.clear();
-                    ProxyConnectionHolder.clear();
+                    ConnectionManager.clear();
                 }
             }
         });
